Remove root command
