--- cmd/root.go
@@ -1,28 +1,0 @@
-package cmd
-
-import (
-	"fmt"
-	"os"
-
-	"github.com/spf13/cobra"
-)
-
-var (
-	verbose bool
-)
-
-var RootCmd = &cobra.Command{
-	Use:   "vedran-daemon",
-	Short: "Vedran daemon cli tool",
-}
-
-func init() {
-	RootCmd.PersistentFlags().BoolVarP(&verbose, "verbose", "v", false, "verbose output")
-}
-
-func Execute() {
-	if err := RootCmd.Execute(); err != nil {
-		fmt.Println(err)
-		os.Exit(1)
-	}
-}
